batch: Move ownership unit validation
Ownership unit construction now has separate selection and rebuild helpers,
while callers still validate completed units before applying or previewing
ownership changes.

Validation still lives in the construction module. That keeps
completed-unit safety checks next to display grouping even though
validation only consumes unit records and reports malformed metadata.

This commit continues separating ownership unit responsibilities by moving
structural validation into an ownership unit validation module and updating
source callers to import that boundary.

Subsequent commits will record validation ownership in the architecture
tests.
